Refactor : 기존 repository 폴더 model + repository로 분기
