+++ src/main/kotlin/com/everywaffle/team3server/post/repository/PostLikeRepository.kt
@@ -1,5 +1,6 @@
 package com.everywaffle.team3server.post.repository
 
+import com.everywaffle.team3server.post.model.PostLikeEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface PostLikeRepository : JpaRepository<PostLikeEntity, Long>
