+++ src/main/kotlin/com/everywaffle/team3server/post/repository/PostRepository.kt
@@ -1,5 +1,6 @@
 package com.everywaffle.team3server.post.repository
 
+import com.everywaffle.team3server.post.model.PostEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface PostRepository : JpaRepository<PostEntity, Long>
